Defect 117632 - Run bar delete button needs to confirm

+++ bundles/org.eclipse.orion.client.ui/web/orion/projectCommands.js
@@ -300,17 +300,21 @@ define(['require', 'i18n!orion/navigate/nls/messages', 'orion/webui/littlelib', 
 	}
 
 	function runDeleteLaunchConfiguration(launchConf, context){
-		var deferreds = [];
-		deferreds.push(context.projectClient.deleteProjectLaunchConfiguration(launchConf)); /* deletes the launch configuration file if present */
-		deferreds.push(context.deployService.deleteAdditionalLaunchConfiguration(context.project, launchConf)); /* deletes additional launch configuration data if present */
-
-		progress.showWhile(Deferred.all(deferreds), messages["deletingLaunchConfiguration"], true).then(function(){
-			fileDispatcher.dispatchEvent({
-				type: "delete",
-				oldValue: launchConf.File, /* TODO: Find out what happens when there's no File */
-				parent: launchConf.File.parent
-			});
-		}, context.errorHandler);
+		var msg = i18nUtil.formatMessage(messages["confirmLaunchDelete"], launchConf.Name);
+		context.commandService.confirm(null, msg, messages.OK, messages.Cancel, true, function(doit) {
+			if (!doit) return;
+			var deferreds = [];
+			deferreds.push(context.projectClient.deleteProjectLaunchConfiguration(launchConf)); /* deletes the launch configuration file if present */
+			deferreds.push(context.deployService.deleteAdditionalLaunchConfiguration(context.project, launchConf)); /* deletes additional launch configuration data if present */
+			
+			progress.showWhile(Deferred.all(deferreds), messages["deletingLaunchConfiguration"], true).then(function(){
+				fileDispatcher.dispatchEvent({
+					type: "delete",
+					oldValue: launchConf.File, /* TODO: Find out what happens when there's no File */
+					parent: launchConf.File.parent
+				});
+			}, context.errorHandler);
+		});
 	}
 
 	var sharedDependencyDispatcher;
